requirements.txt
requests>=2.31.0
click>=8.1.0
openpyxl>=3.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0
